+++ CHANGELOG.rst
@@ -1,13 +1,15 @@
 Change Log
 ==========
 
-Unreleased
-----------
+0.2.2 (2017-07-22)
+------------------
 
 - Add ``documenteer.sphinxext.bibtex`` extension to support LSST BibTeX entries that include a ``docushare`` field.
   Originally from `lsst-texmf`_.
+  This extension is active in the technote Sphinx configuration.
 - Add a ``refresh-lsst-bib`` command line program that downloads the latest LSST bib files from the `lsst-texmf`_ GitHub repository.
   This program can be used by technote authors to update a technote's local bibliography set at any time.
+- Added graceful defaults when a technote is being built without an underlying Git repository (catches exceptions from functions that seek Git metadata).
 - Add a dependency upon the Requests library.
 
 0.2.1 (2017-07-21)